requirements.txt
langchain>=0.3.0
langgraph>=0.2.35
langsmith>=0.1.77
langchain-community>=0.3.0
langchain-groq>=0.1.3
sentence-transformers>=3.1.1
qdrant-client>=1.8.2
pypdf>=4.3.1
python-dotenv>=1.0.1
streamlit>=1.39.0
requests>=2.32.3
pytest>=8.2.1
typing_extensions>=4.12.2
torch>=2.1.0
langchain-qdrant
